feat: change the structure of the boilerplate
