--- src/app/api.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ApiService } from './services/api.service';
-
-describe('ApiService', () => {
-  let service: ApiService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ApiService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
