+++ src/app/app.component.html
@@ -1,8 +1,2 @@
-<main class="section__principal">
-  <nav class="navbar__container">
-    <a routerLinkActive="active"  [routerLinkActiveOptions]="{exact: true}" routerLink="/">Inicio</a>
-    <a routerLinkActive="active" routerLink="/templates">Plantillas</a>
-  </nav>
-</main>
 
 <router-outlet />
