--- src/app/home/home.component.html
@@ -1,1 +1,0 @@
-<h1>Bienvenido</h1>
