+++ src/app/modules/auth/auth.routes.ts
@@ -1,0 +1,11 @@
+import { Routes } from "@angular/router";
+import { LoginComponent } from "./login/login.component";
+
+const routes: Routes = [
+  {
+    path: 'login',
+    component: LoginComponent,
+  }
+];
+
+export default routes;
