+++ src/app/modules/home/home-page/home-page.component.css
@@ -1,0 +1,5 @@
+/* .principal__home--section{
+    border: 1px solid yellow;
+    height: 100vh;
+  
+} */
