+++ src/app/modules/home/home-page/home-page.component.html
@@ -1,0 +1,7 @@
+<div class="principal__home--section">
+  <app-nav-bar></app-nav-bar>
+  <app-side-bar></app-side-bar>
+  <div>
+    <router-outlet></router-outlet>
+  </div>
+</div>
