+++ src/app/modules/home/welcome/welcome.component.html
@@ -1,0 +1,4 @@
+<section class="welcome__section">
+  <h1>welcome works!</h1>
+  <a class="btn__sesion" [routerLink]="['/auth/login']"> Inicia sesion </a>
+</section>
