+++ src/app/modules/home/welcome/welcome.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { RouterLink } from '@angular/router';
+
+@Component({
+  selector: 'app-welcome',
+  standalone: true,
+  imports: [RouterLink],
+  templateUrl: './welcome.component.html',
+  styleUrl: './welcome.component.css'
+})
+export class WelcomeComponent {
+
+}
