+++ src/app/modules/templates/details-template/details-template.component.css
