+++ src/app/shared/components/nav-bar/nav-bar.component.html
@@ -1,0 +1,5 @@
+<main class="section__principal">
+  <nav class="navbar__container">
+    navbar
+  </nav>
+</main>
