--- src/app/table-templates/table-templates.component.css
@@ -1,83 +1,0 @@
-.tableTemplate__container{
-  display: flex;
-  flex-direction: column;
-  text-align: center;
-  align-items: center;
-
-  h1{
-    color: var(--green-principal);
-  }
-
-  table {
-    border-collapse: collapse;
-    color: white;
-    font-weight: bold;
-    height: 80px;
-    max-width: 1440px;
-    margin: 20px 0;
-  
-    thead {
-      border: 2px solid white;
-      background-color: #0b1327;
-      color: var(--green-principal);
-        
-    }
-
-    th {
-      text-align: center;
-      padding: 20px;
-    }
-    
-    tbody{
-      border: 2px solid white;
-
-      a{
-        text-decoration: none;
-        color: white;
-        
-        &:hover{
-          cursor: pointer;
-          color: var(--disabled-green);
-        }
-      }
-
-      td {
-        padding: 20px;
-        text-align: left;
-        border: 2px solid white;
-        
-      }
-
-      tr:nth-child(odd) {
-          color: white;
-      }
-      
-      tr:nth-child(even) {
-          background-color: #0b1327;
-      }
-    }
-
-    .btn__delete{
-      background-color: var(--delete-red);
-      color: white;
-      margin: 2px 5px;
-      
-      &:hover{
-        background-color: var(--hot-red);
-      }
-    }
-
-    .btn__edit{
-      background-color: var(--disabled-green);
-      color: white;
-      margin: 2px 5px;
-
-      &:hover{
-        background-color: var(--green-principal);
-      }
-    }
-    
-  }
-
-}
-
